pyproject.toml
[project]
name = "bound-explain"
version = "0.0.1"
description = "Boundary explainability pipeline built on SigLIP2 activations."
authors = ["Codex <codex@example.com>"]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.1",
    "torchvision",
    "numpy",
    "h5py",
    "pandas",
    "scikit-learn",
    "tqdm",
    "Pillow",
    "matplotlib",
    "transformers",
    "scipy",
]

[build-system]
requires = ["setuptools>=65.5.0", "wheel"]
build-backend = "setuptools.build_meta"
